full_stack_app/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-adk>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.27.0
ag-ui-adk>=0.1.0
